xpos float32

+++ attention.py
@@ -80,7 +80,7 @@ class BahdanauAttention(tf.keras.Model):
                                 lambda: tf.concat([att_value, values_slice], axis=1))
             
             # get x asix numbers from start to end
-            xranges = tf.range(start, end, dtype=tf.float32)
+            xranges = tf.cast(tf.range(start, end), dtype=tf.float32)
             xranges = tf.cond(tf.equal(xlen, self.att_len), 
                                 lambda: xranges,
                                 lambda: tf.concat([xranges, tf.zeros([self.att_len-xlen], dtype=tf.float32)], axis=0))
@@ -94,7 +94,7 @@ class BahdanauAttention(tf.keras.Model):
 
         zero = lambda: tf.constant(0, dtype=tf.int32)
         zeros= lambda: tf.zeros([self.att_len, 1, self.units], dtype=tf.float32)
-        zeros2= lambda: tf.zeros([self.batch_sz], dtype=tf.float32)
+        zeros2= lambda: tf.zeros([self.batch_sz], dtype=tf.int32)
         i = tf.Variable(initial_value=zero, dtype=tf.int32)
         att_value = tf.Variable(initial_value=zeros, dtype=tf.float32)
         xpos = tf.Variable(initial_value=zeros2, dtype=tf.float32)
